Code fixes

+++ ApplicationLibCode/GeoMech/GeoMechDataModel/RigFemPart.cpp
@@ -146,7 +146,7 @@ bool RigFemPart::fillElementCoordinates( size_t elementIdx, std::array<cvf::Vec3
 
     // Fill coordinates for each node
     const auto& partNodes = nodes();
-    for ( int i = 0; i < nodeIndices.size(); ++i )
+    for ( int i = 0; i < (int)nodeIndices.size(); ++i )
     {
         coordinates[i].set( partNodes.coordinates[nodeIndices[i]] );
     }
@@ -661,6 +661,6 @@ std::vector<std::string> RigFemPart::elementSetNames() const
 //--------------------------------------------------------------------------------------------------
 std::vector<size_t> RigFemPart::elementSet( int setIndex ) const
 {
-    if ( ( setIndex < 0 ) || ( setIndex >= m_elementIndexSets.size() ) ) return {};
+    if ( ( setIndex < 0 ) || ( setIndex >= (int)m_elementIndexSets.size() ) ) return {};
     return m_elementIndexSets[setIndex];
 }
